Darken CursorLine, add ShowTrailingWhitespace
If you use the CursorLine, it was the same color as the StatusLine, so
I've darkened it just a hair.  I've also added something from my own
init.vim: Support for vim-scripts/ShowTrailingWhitespace.  This is such
a handy thing for anyone working with DVCS tools like git!  It also
helps for things like Markdown where a trailing space can be syntactic.